{{ cookiecutter.project_slug }}/code/requirements.txt
# neuroml bits
pyNeuroML

# visualisation
pyNeuroML[vispy]

# validation
osbmodelvalidation

# simulation engine
{% if cookiecutter.simulation_engine|lower != "jlems" -%}
pyNeuroML[{{ cookiecutter.simulation_engine }}] {% endif -%}
